backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.3.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.1.4
django-filter==23.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.4.2
mysqlclient==2.2.7
Pillow==10.1.0
prompt_toolkit==3.0.51
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
